+++ src/egeo-demo/public_api.ts
@@ -19,10 +19,10 @@ export { StCheckboxDemoModule } from './st-checkbox-demo/st-checkbox-demo.module
 export { StDropdownMenuDemoModule } from './st-dropdown-menu-demo/st-dropdown-menu-demo.module';
 export { StFooterDemoModule } from './st-footer-demo/st-footer-demo.module';
 export { StFormDemoModule } from './st-form-demo/st-form-demo.module';
+export { StFormListDemoModule } from './st-form-list-demo/st-form-list-demo.module';
 export { StFullscreenLayoutDemoModule } from './st-fullscreen-layout-demo/st-fullscreen-layout-demo.module';
 export { StHeaderDemoModule } from './st-header-demo/st-header-demo.module';
 export { StHelpDemoModule } from './st-help-demo/st-help-demo.module';
-import { StHorizontalTabsDemoModule } from './st-horizontal-tabs-demo/st-horizontal-tabs-demo.module';
 export { StInfoBoxDemoModule } from './st-info-box-demo/st-info-box-demo.module';
 export { StInfoCardDemoModule } from './st-info-card-demo/st-info-card-demo.module';
 export { StInputDemoModule } from './st-input-demo/st-input-demo.module';