+++ src/lib/public_api.ts
@@ -87,6 +87,8 @@ export { StHelpModule } from './st-help/st-help.module';
 
 // Dynamic form
 export { StFormModule } from './st-form/st-form.module';
+export { StFormFieldModule } from './st-form/st-form-field/st-form-field.module';
+export { StFormListModule } from './st-form-list/st-form-list.module';
 
 // Fullscreen layout
 export { StFullscreenLayoutModule } from './st-fullscreen-layout/st-fullscreen-layout.module';