+++ src/lib/st-form/st-form.component.ts
@@ -8,33 +8,102 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { Component, Input, OnInit, ChangeDetectionStrategy } from '@angular/core';
-import { FormGroup, FormControl, FormArray } from '@angular/forms';
+import {
+   Component,
+   Input,
+   OnInit,
+   ChangeDetectionStrategy,
+   forwardRef,
+   ChangeDetectorRef,
+   ViewChild
+} from '@angular/core';
+import {
+   ControlValueAccessor, NG_VALUE_ACCESSOR, FormGroup, FormControl, NgForm
+} from '@angular/forms';
 
 @Component({
    selector: 'st-form',
    templateUrl: './st-form.component.html',
-   changeDetection: ChangeDetectionStrategy.OnPush
+   changeDetection: ChangeDetectionStrategy.OnPush,
+   providers: [
+      { provide: NG_VALUE_ACCESSOR, useExisting: forwardRef(() => StFormComponent), multi: true }
+   ]
 })
 
-export class StFormComponent implements OnInit {
+export class StFormComponent implements ControlValueAccessor, OnInit {
    @Input() schema: any;
-   @Input() form: FormGroup = new FormGroup({});
-   @Input() model: any = {};
+   @Input() name: string;
+   @ViewChild('form') form: NgForm;
+
+   public formGroup: FormGroup = new FormGroup({});
+
+   private _value: any = {};
+
+   @Input()
+   get value(): any {
+      return this._value;
+   }
+
+   set value(value: any) {
+      if (value && value !== this._value) {
+         this._value = value;
+         this.onChange(this.value);
+         this._cd.markForCheck();
+      }
+   }
+
+
+   // Function to call when the value changes.
+   onChange(_: any): void {
+      this._cd.markForCheck();
+   }
+
+   onTouched = () => {
+   }
+
+   constructor(private _cd: ChangeDetectorRef) {
+   }
 
    ngOnInit(): void {
       Object.keys(this.schema.properties).forEach(propertyName => {
          let property: any = this.schema.properties[propertyName];
-         let formControl: FormControl | FormArray;
-         if (property.default && this.model[propertyName] === undefined) {
-            this.model[propertyName] = property.default;
+         if (property.default && this._value[propertyName] === undefined) {
+            this._value[propertyName] = property.default;
          }
-         formControl = new FormControl(this.model[propertyName] || '');
-         this.form.addControl(propertyName, formControl);
+         this.formGroup.addControl(propertyName, new FormControl(this._value[propertyName] || property.default || null));
       });
    }
 
    isRequired(propertyName: string): boolean {
       return propertyName && this.schema.required && this.schema.required.indexOf(propertyName) !== -1;
    }
+
+   // When value is received from outside
+   writeValue(value: any): void {
+      this._value = value;
+      this.onChange(value);
+      this._cd.markForCheck();
+   }
+
+   onChangeProperty(value: any, property: string): void {
+      this._value[property] = value;
+      this.onChange(this.value);
+      this._cd.markForCheck();
+
+   }
+
+   // Registry the change function to propagate internal model changes
+   registerOnChange(fn: (_: any) => void): void {
+      this.onChange = fn;
+   }
+
+   // Registry the touch function to propagate internal touch events TODO: make this function.
+   registerOnTouched(fn: () => void): void {
+      this.onTouched = fn;
+   }
+
+   // Allows Angular to disable the list.
+   setDisabledState(isDisabled: boolean): void {
+   }
+
 }