+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## 8.2.0 (upcoming)
 
+**New features:**
+
+* st-form-list: Allow to generate list of items dynamically
+
 **Fixed bugs:**
 
 * st-header: Fix z-index 
-
 
 
 ## 8.1.1 (February 02, 2018)